refactor: move `Lexer` to its own file
